NO-TICKET: reposition {auction,mint,proof_of_stake,standard_payment}
