+++ types/src/auction/mod.rs
@@ -6,7 +6,6 @@ mod detail;
 mod era_info;
 mod providers;
 mod seigniorage_recipient;
-mod types;
 mod unbonding_purse;
 
 use alloc::{collections::BTreeMap, vec::Vec};
@@ -27,8 +26,31 @@ pub use providers::{
     AccountProvider, MintProvider, RuntimeProvider, StorageProvider, SystemProvider,
 };
 pub use seigniorage_recipient::SeigniorageRecipient;
-pub use types::*;
 pub use unbonding_purse::UnbondingPurse;
+
+/// Representation of delegation rate of tokens. Fraction of 1 in trillionths (12 decimal places).
+pub type DelegationRate = u64;
+
+/// Validators mapped to their bids.
+pub type Bids = BTreeMap<PublicKey, Bid>;
+
+/// Weights of validators. "Weight" in this context means a sum of their stakes.
+pub type ValidatorWeights = BTreeMap<PublicKey, U512>;
+
+/// Era index type.
+pub type EraId = u64;
+
+/// List of era validators
+pub type EraValidators = BTreeMap<EraId, ValidatorWeights>;
+
+/// Collection of seigniorage recipients.
+pub type SeigniorageRecipients = BTreeMap<PublicKey, SeigniorageRecipient>;
+
+/// Snapshot of `SeigniorageRecipients` for a given era.
+pub type SeigniorageRecipientsSnapshot = BTreeMap<EraId, SeigniorageRecipients>;
+
+/// Validators and delegators mapped to their unbonding purses.
+pub type UnbondingPurses = BTreeMap<PublicKey, Vec<UnbondingPurse>>;
 
 /// Bonding auction contract interface
 pub trait Auction:
